OSX support via the Apple Core Audio, Audio Unit C API. Only supports f32 so far.

+++ src/null/mod.rs
@@ -18,7 +18,7 @@ impl Voice {
         ::SampleFormat::U16
     }
 
-    pub fn append_data<'a, T>(&'a mut self, _: uint) -> Buffer<'a, T> {
+    pub fn append_data<'a, T>(&'a mut self, _: usize) -> Buffer<'a, T> {
         Buffer
     }
